+++ lib/main.dart
@@ -34,7 +34,7 @@ class MyApp extends StatelessWidget {
           theme: ThemeData(
             primarySwatch: Colors.blue,
           ),
-         home: EventList(),
+         home: ListScreen(),
        
         );
       },
